Update transactions processor
