+++ services/blockchain-indexer/shared/indexer/transactionProcessor/interoperability/registerSidechain.js
@@ -34,16 +34,15 @@ const getBlockchainAppsTable = () => getTableInstance(
 );
 
 // Command specific constants
-const COMMAND_NAME = 'mainchainRegistration';
+const COMMAND_NAME = 'registerSidechain';
 
-// TODO: Needs work
 const applyTransaction = async (blockHeader, tx, dbTrx) => {
 	const blockchainAppsTable = await getBlockchainAppsTable();
 
-	logger.trace(`Indexing mainchain (${tx.params.chainID}) registration information.`);
+	logger.trace(`Indexing sidechain (${tx.params.chainID}) registration information.`);
 	const appInfo = {
-		chainID: tx.params.ownChainID,
-		name: tx.params.ownName,
+		chainID: tx.params.chainID,
+		name: tx.params.name,
 		state: '', // TODO: Set init state from events
 		address: getLisk32AddressFromPublicKey(tx.senderPublicKey),
 		lastUpdated: blockHeader.timestamp,
@@ -51,16 +50,15 @@ const applyTransaction = async (blockHeader, tx, dbTrx) => {
 	};
 
 	await blockchainAppsTable.upsert(appInfo, dbTrx);
-	logger.debug(`Indexed mainchain (${tx.params.chainID}) registration information.`);
+	logger.debug(`Indexed sidechain (${tx.params.chainID}) registration information.`);
 };
 
-// eslint-disable-next-line no-unused-vars
 const revertTransaction = async (blockHeader, tx, dbTrx) => {
 	const blockchainAppsTable = await getBlockchainAppsTable();
 
-	logger.trace(`Reverting mainchain (${tx.params.chainID}) registration information.`);
+	logger.trace(`Reverting sidechain (${tx.params.chainID}) registration information.`);
 	await blockchainAppsTable.deleteByPrimaryKey(tx.params.chainID, dbTrx);
-	logger.debug(`Reverted mainchain (${tx.params.chainID}) registration information.`);
+	logger.debug(`Reverted sidechain (${tx.params.chainID}) registration information.`);
 };
 
 module.exports = {
